Remove all 'PageSwitch' from the back buttons and use universal back button code 0x2295 where possible
This allows the back button to be controlled from a single place in the firmware.
Also by removing the HasPageSwitch from the buttons we guarantee that the firmware controls which page is displayed - so that if an synchronous operation happens the user can't switch to a different page without the firmware knowing about it.