Adds archived (7z) files containing the DGUS tools
DGUS Toolv8.0 was used to build and maintain versions 5 and 6 of the Community display firmware.

DGUS Tool v8.2 will be used to build and maintain the Community display firmware after version 6.1-Final.  It is too large to store as a single file in GitHub, so it has been split into two files.  Download and recombine those two files with 7zip.