Update note on Screen 78
The note says assumes a hot-end capable of 6mm retraction, but the code actually commands 4mm of retraction.  A future release will enable users to modify this value by editing the value of G26_RETRACT_MULTIPLIER in Configuration.h.
At that point, this note should be reviewed and probably should be revised to whatever value is baked into the released compiled firmware.