Uploading latest nextgen UI firmware
The LED ON/OFF button on the Printing->Tune menu has now been nudged to the correct location.
The M117 Scrolling Text field is not displaying messages on the screen, but I have not yet deduced the why not.
One way to solve the problem may be to ensure that all messages are less than 29 characters in length, which will then be displayed to the M117 Static text field.