Add caution to remnant from previous readme - being reworked and replaced
When the refactored UI PR was merged into extui, a large amount of useful but outdated info was removed.  This commit preserves that info but renames the file and removes some front-matter that no longer belongs in there.
Thinkersbluff is presently reorganizing and updating this info into a series of short documents, which may fit best into a Wiki format. (TBD).