Changed Z-Offset adjust buttons from Constant to Disposable setting types
Eliminates a bug I had not yet discovered, with thanks to InsanityAutomation for the heads-up and the fix.  Did not change the version# since is still in WIP, and finding the discipline of updating the two bitmaps every time is a pest.
Think I need to change the way I do that branding...