Restructured and Clarified the Installation Guidelines and Screen Images
At Release 1.1, some of the documentation remained confusing and some of the imagery was out-dated + not correctly linked into the documents.

Rather than correctly re-label the release to 1.1.1, I am taking the unusual and controversial approach os simply correcting the source files and updating the release with a new zipfile.  This will mean the Source.zip files attached to the Release are NOT updated (see the src folder in the repo instead), but the DWIN_SET .zip file is correct.

This seems like the lesser of two evils, rather than trying to communicate the new revision # when no changes were actually made to the firmware...