Delete DGUS_Tool_v7641-0801.zip
Thought this would be the right toolset to use, as it is the latest online at DWIN.
Turns-out, I was wrong.
Still need to use DGUSTool V8.2.1.14 to generate DWIN_SETs compatible with all versions of DGUS2.
Might as well reduce the size of the repo, and help others avoid my mistake.