Update Flashing Help file and images to include flashing to DGUS2 v4.5
The flashing process is the same for both versions of DGUS2, but the screens displayed during and after flashing are different.  In particular, flashing the CFG file to DGUS2 v4.5 triggers the display of a colourful screen full of noise, which obscures the blue screen you see with DGUS2 v3.5.   This update clarifies that and shows the screens that appear according to which DGUS2 you are flashing to.