Revised document outline. Began authoring new sections
This commit is the first in a series of changes designed to overhaul the development documentation.  The goal is to provide enough information to help the next person who wishes to modify the CR6Community Firmware or to develop their own spinoff.

The term "TBA" throughout this version stands for "To Be Authored"